Вынесли хранение состояний в хуки
